+++ app/routes/_auth.login.jsx
@@ -1,0 +1,18 @@
+function Login() {
+  return (
+    <Login
+      user={res}
+      opened={loginOpen}
+      close={() => {
+        loginToggle()
+        modalToggle()
+      }}
+      signup={() => {
+        loginToggle()
+        signUpToggle()
+      }}
+    />
+  )
+}
+
+export default Login
